modules/batch/docker/requirements.txt
# NumPy 1.x REQUIRED - PyTorch 2.1.0 and Pyannote 3.0.1 were built for NumPy 1.x
# Upgrading to NumPy 2.x will break: pyannote.audio (np.NaN removed)
# To upgrade: PyTorch 2.3+, Pyannote 3.1+
numpy<2

# Supporting libraries (must be installed before torch)
setuptools>=65.0.0
wheel>=0.40.0

# ML Stack - PyTorch 2.1.0 for CUDA 12.1
torch==2.1.0
torchaudio==2.1.0

# AWS SDK
boto3>=1.28.0
awscli>=1.32.0

# Audio Processing - Whisper and Speaker Diarization
# Note: Pyannote 3.0.1 incompatible with NumPy 2.0
faster-whisper==0.10.0
pyannote.audio==3.0.1

